moved privacy request tests from endpoint to dedicated request runner test
