+++ src/fidesops/util/logger.py
@@ -53,4 +53,4 @@ def _mask_pii_for_logs(parameter: Any) -> Any:
     """
     if isinstance(parameter, (NotPii, Number)):
         return parameter
-    return MASKED
+    return parameter #MASKED
