+++ tests/api/v1/endpoints/test_privacy_request_endpoints.py
@@ -2,12 +2,6 @@ import json
 from datetime import datetime
 from typing import List
 from unittest import mock
-from sqlalchemy import (
-    column,
-    table,
-    select,
-)
-import time
 from fastapi_pagination import Params
 import pytest
 from starlette.testclient import TestClient
@@ -25,22 +19,15 @@ from fidesops.api.v1.scope_registry import (
     STORAGE_CREATE_OR_UPDATE,
     PRIVACY_REQUEST_READ,
 )
-from fidesops.db.session import (
-    get_db_engine,
-    get_db_session,
-)
 from fidesops.models.client import ClientDetail
 from fidesops.models.privacy_request import (
     PrivacyRequest,
     ExecutionLog,
     ExecutionLogStatus,
 )
-from fidesops.models.policy import DataCategory, ActionType
+from fidesops.models.policy import ActionType
 from fidesops.schemas.dataset import DryRunDatasetResponse
-from fidesops.service.connectors import PostgreSQLConnector
 from fidesops.util.cache import get_identity_cache_key, get_encryption_cache_key
-from ....test_support import wait_for_privacy_request, wait_for
-import time
 
 page_size = Params().size
 
@@ -296,285 +283,6 @@ class TestCreatePrivacyRequest:
         response_data = resp.json()["failed"]
         assert len(response_data) == 1
 
-    @pytest.mark.integration
-    def test_create_and_process_access_request(
-        self,
-        postgres_example_test_dataset_config_read_access,
-        url,
-        db,
-        api_client: TestClient,
-        generate_auth_header,
-        policy,
-    ):
-        customer_email = "customer-1@example.com"
-        data = [
-            {
-                "requested_at": "2021-08-30T16:09:37.359Z",
-                "policy_key": policy.key,
-                "identities": [{"email": customer_email}],
-            }
-        ]
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_CREATE])
-        resp = api_client.post(url, json=data, headers=auth_header)
-        assert resp.status_code == 200
-        response_data = resp.json()["succeeded"]
-        wait_for_privacy_request(db, response_data[0]["id"])
-        time.sleep(2)
-        assert len(response_data) == 1
-        pr = PrivacyRequest.get(db=db, id=response_data[0]["id"])
-        results = pr.get_results()
-        assert len(results.keys()) == 11
-
-        for key in results.keys():
-            assert results[key] is not None
-            assert results[key] != {}
-
-        result_key_prefix = (
-            f"EN_{pr.id}__access_request__postgres_example_test_dataset:"
-        )
-        customer_key = result_key_prefix + "customer"
-        assert results[customer_key][0]["email"] == customer_email
-
-        visit_key = result_key_prefix + "visit"
-        assert results[visit_key][0]["email"] == customer_email
-        log_id = pr.execution_logs[0].id
-        pr_id = pr.id
-
-        policy.delete(db=db)
-        pr.delete(db=db)
-        db.expunge_all()
-        assert ExecutionLog.get(db, id=log_id).privacy_request_id == pr_id
-
-    @pytest.mark.integration_erasure
-    def test_create_and_process_erasure_request_specific_category(
-        self,
-        postgres_example_test_dataset_config,
-        url,
-        db,
-        api_client: TestClient,
-        generate_auth_header,
-        erasure_policy,
-        read_connection_config,
-    ):
-        customer_email = "customer-1@example.com"
-        customer_id = 1
-        data = [
-            {
-                "requested_at": "2021-08-30T16:09:37.359Z",
-                "policy_key": erasure_policy.key,
-                "identities": [{"email": customer_email}],
-            }
-        ]
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_CREATE])
-        resp = api_client.post(url, json=data, headers=auth_header)
-
-        assert resp.status_code == 200
-        response_data = resp.json()["succeeded"]
-        assert len(response_data) == 1
-        wait_for_privacy_request(db, response_data[0]["id"])
-        time.sleep(2)
-        pr = PrivacyRequest.get(db=db, id=response_data[0]["id"])
-        pr.delete(db=db)
-
-        example_postgres_uri = PostgreSQLConnector(read_connection_config).build_uri()
-        engine = get_db_engine(database_uri=example_postgres_uri)
-        SessionLocal = get_db_session(engine=engine)
-        integration_db = SessionLocal()
-        stmt = select(
-            column("id"),
-            column("name"),
-        ).select_from(table("customer"))
-        res = integration_db.execute(stmt).all()
-
-        customer_found = False
-        for row in res:
-            if customer_id in row:
-                customer_found = True
-                # Check that the `name` field is `None`
-                assert row[1] is None
-        assert customer_found
-
-    @pytest.mark.integration_erasure
-    def test_create_and_process_erasure_request_generic_category(
-        self,
-        postgres_example_test_dataset_config,
-        url,
-        db,
-        api_client: TestClient,
-        generate_auth_header,
-        erasure_policy,
-        connection_config,
-    ):
-        # It's safe to change this here since the `erasure_policy` fixture is scoped
-        # at function level
-        target = erasure_policy.rules[0].targets[0]
-        target.data_category = DataCategory("user.provided.identifiable.contact").value
-        target.save(db=db)
-
-        email = "customer-2@example.com"
-        customer_id = 2
-        data = [
-            {
-                "requested_at": "2021-08-30T16:09:37.359Z",
-                "policy_key": erasure_policy.key,
-                "identities": [{"email": email}],
-            }
-        ]
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_CREATE])
-        resp = api_client.post(url, json=data, headers=auth_header)
-
-        assert resp.status_code == 200
-        response_data = resp.json()["succeeded"]
-        assert len(response_data) == 1
-        wait_for_privacy_request(db, response_data[0]["id"])
-        time.sleep(2)
-        pr = PrivacyRequest.get(db=db, id=response_data[0]["id"])
-        pr.delete(db=db)
-
-        example_postgres_uri = PostgreSQLConnector(connection_config).build_uri()
-        engine = get_db_engine(database_uri=example_postgres_uri)
-        SessionLocal = get_db_session(engine=engine)
-        integration_db = SessionLocal()
-        stmt = select(
-            column("id"),
-            column("email"),
-            column("name"),
-        ).select_from(table("customer"))
-        res = integration_db.execute(stmt).all()
-
-        customer_found = False
-        for row in res:
-            if customer_id in row:
-                customer_found = True
-                # Check that the `email` field is `None` and that its data category
-                # ("user.provided.identifiable.contact.email") has been erased by the parent
-                # category ("user.provided.identifiable.contact")
-                assert row[1] is None
-                assert row[2] is not None
-            else:
-                # There are two rows other rows, and they should not have been erased
-                assert row[1] in ["customer-1@example.com", "jane@example.com"]
-        assert customer_found
-
-    @pytest.mark.integration_erasure
-    def test_create_and_process_erasure_request_with_table_joins(
-        self,
-        postgres_example_test_dataset_config,
-        url,
-        db,
-        api_client: TestClient,
-        generate_auth_header,
-        erasure_policy,
-        connection_config,
-    ):
-        # It's safe to change this here since the `erasure_policy` fixture is scoped
-        # at function level
-        target = erasure_policy.rules[0].targets[0]
-        target.data_category = DataCategory(
-            "user.provided.identifiable.financial"
-        ).value
-        target.save(db=db)
-
-        customer_email = "customer-1@example.com"
-        customer_id = 1
-        data = [
-            {
-                "requested_at": "2021-08-30T16:09:37.359Z",
-                "policy_key": erasure_policy.key,
-                "identities": [{"email": customer_email}],
-            }
-        ]
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_CREATE])
-        resp = api_client.post(url, json=data, headers=auth_header)
-
-        assert resp.status_code == 200
-        response_data = resp.json()["succeeded"]
-        assert len(response_data) == 1
-        wait_for_privacy_request(db, response_data[0]["id"])
-        time.sleep(2)
-        pr = PrivacyRequest.get(db=db, id=response_data[0]["id"])
-        pr.delete(db=db)
-
-        example_postgres_uri = PostgreSQLConnector(connection_config).build_uri()
-        engine = get_db_engine(database_uri=example_postgres_uri)
-        SessionLocal = get_db_session(engine=engine)
-        integration_db = SessionLocal()
-        stmt = select(
-            column("customer_id"),
-            column("id"),
-            column("ccn"),
-            column("code"),
-            column("name"),
-        ).select_from(table("payment_card"))
-        res = integration_db.execute(stmt).all()
-
-        card_found = False
-        for row in res:
-            if row[0] == customer_id:
-                card_found = True
-                assert row[2] is None
-                assert row[3] is None
-                assert row[4] is None
-
-        assert card_found is True
-
-    @pytest.mark.integration_erasure
-    def test_create_and_process_erasure_request_read_access(
-        self,
-        postgres_example_test_dataset_config_read_access,
-        url,
-        db,
-        api_client: TestClient,
-        generate_auth_header,
-        erasure_policy,
-        connection_config,
-    ):
-        customer_email = "customer-2@example.com"
-        customer_id = 2
-        data = [
-            {
-                "requested_at": "2021-08-30T16:09:37.359Z",
-                "policy_key": erasure_policy.key,
-                "identities": [{"email": customer_email}],
-            }
-        ]
-        auth_header = generate_auth_header(scopes=[PRIVACY_REQUEST_CREATE])
-        resp = api_client.post(url, json=data, headers=auth_header)
-
-        assert resp.status_code == 200
-        response_data = resp.json()["succeeded"]
-        assert len(response_data) == 1
-        wait_for_privacy_request(db, response_data[0]["id"])
-        time.sleep(1)
-        pr = PrivacyRequest.get(db=db, id=response_data[0]["id"])
-        errored_execution_logs = pr.execution_logs.filter_by(status="error")
-        assert errored_execution_logs.count() == 9
-        assert (
-            errored_execution_logs[0].message
-            == "No values were erased since this connection "
-            "my-postgres-db-1-read-config has not been given write access"
-        )
-        pr.delete(db=db)
-
-        example_postgres_uri = PostgreSQLConnector(connection_config).build_uri()
-        engine = get_db_engine(database_uri=example_postgres_uri)
-        SessionLocal = get_db_session(engine=engine)
-        integration_db = SessionLocal()
-        stmt = select(
-            column("id"),
-            column("name"),
-        ).select_from(table("customer"))
-        res = integration_db.execute(stmt).all()
-
-        customer_found = False
-        for row in res:
-            if customer_id in row:
-                customer_found = True
-                # Check that the `name` field is NOT `None`. We couldn't erase, because the ConnectionConfig only had
-                # "read" access
-                assert row[1] is not None
-        assert customer_found
-
 
 class TestGetPrivacyRequests:
     @pytest.fixture(scope="function")
@@ -801,7 +509,7 @@ class TestGetPrivacyRequests:
         assert (
             postgres_execution_log.updated_at < second_postgres_execution_log.updated_at
         )
-        wait_for_privacy_request(db, privacy_request.id)
+        #wait_for_privacy_request(db, privacy_request.id)
         expected_resp = {
             "items": [
                 {
